requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
langchain==0.0.346
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
PyPDF2==3.0.1
gtts==2.3.2
moviepy==1.0.3
pillow==10.0.1
python-multipart==0.0.6